audiosource/wavstream: Add an mpg123 decoder backend, preferred over dr_mp3.
This circumvents rare edge-case (and non-overlapping!) issues in both the dr_mp3 and ffmpeg decoders, on some pesky mp3s.

Putting this here as a reference for if I ever get around to reporting this issue to dr_mp3:
https://osu.ppy.sh/beatmapsets/545920
This mp3 plays with the wrong offset (with dr_mp3), but the duration is calculated correctly.
Neither mpg123 and ffmpeg have this issue.

Also note: this mp3 has +7ms duration with ffmpeg (dr_mp3 is fine here), but that's probably not an ffmpeg bug and more PEBKAC.
https://osu.ppy.sh/beatmapsets/2252345