backends/miniaudio: Graceful recovery for audio device/server interruptions or changes.
TODO: This doesn't work in the case of restarting the (pipewire-)pulse server, a full deinit->init cycle needs to manually be performed.
      It at least unblocks that, though. Before this, deinit() blocked forever since miniaudio tried to unregister a pulse context from a server that no longer existed.
      I think it works as expected for device changes, though (on Windows and Linux).

TODO (2): Reinitializing on device changes can permanently increase the buffer size, if the device is changed to a higher latency device then back to a low latency one.

TODO (3): It's overly complicated. Tons of state, mutexes, and atomic variables which shouldn't be required. Will be refactored at some point.