audiosource/wavstream: Fix more ogg decoding issues.
mOffset was double-counted, and getOggData didn't properly handle misaligned frames while seeking.
Replace it with a temp-buffered approach similar to the other decoders.

Also, added a minor miniaudio tweak to set the "Pro Audio" device profile for WASAPI. This may or may not do anything.

And decreased version number back to 1.1.7, because I never made a tag for that...